CeleryScript Namespace
