+++ lib/celery_script/tree_climber.rb
@@ -1,0 +1,14 @@
+module CeleryScript
+  class TreeClimber
+    def self.each_node(node, &blk)
+    end
+
+  private
+
+    def self.traverse(node, &blk)
+      if node.is_a?(AstNode)
+      else
+      end
+    end
+  end
+end
